infra/cloudcustodian/requirements.txt
aws-cdk-lib==2.132.1
constructs>=10.0.0,<11.0.0
c7n==0.9.35